+++ audio_conversion/test/AudioConversionTest.cpp
@@ -60,11 +60,11 @@ TEST_P(AudioConversionT, audioConversion)
 
     const void *sourceBuf = std::tr1::get<2>(GetParam());
     size_t sourceBufSize = std::tr1::get<3>(GetParam());
-    const uint32_t inputFrames = sampleSpecSrc.convertBytesToFrames(sourceBufSize);
+    const size_t inputFrames = sampleSpecSrc.convertBytesToFrames(sourceBufSize);
 
     const uint8_t *expectedDstBuf = (uint8_t *)std::tr1::get<4>(GetParam());
     size_t expectedDstBufSize = std::tr1::get<5>(GetParam());
-    const uint32_t expextedDstFrames = AudioUtils::convertSrcToDstInFrames(inputFrames,
+    const size_t expextedDstFrames = AudioUtils::convertSrcToDstInFrames(inputFrames,
                                                                            sampleSpecSrc,
                                                                            sampleSpecDst);
 
@@ -75,11 +75,11 @@ TEST_P(AudioConversionT, audioConversion)
 
     bool allocateBuffer = std::tr1::get<6>(GetParam());
     uint8_t *dstBuf = NULL;
-    uint32_t dstFrames = 0;
+    size_t dstFrames = 0;
 
     if (allocateBuffer) {
 
-        uint32_t dstSizeInBytes = AudioUtils::convertSrcToDstInBytes(sourceBufSize,
+        size_t dstSizeInBytes = AudioUtils::convertSrcToDstInBytes(sourceBufSize,
                                                                      sampleSpecSrc,
                                                                      sampleSpecDst);
 
@@ -612,11 +612,11 @@ TEST(AudioConversion, resampleDoubleConfigure)
         0xFFFF, 0xFFFF,
         0xF000, 0x9880
     };
-    const uint32_t inputFrames =
+    const size_t inputFrames =
         sizeof(sourceBuf) / (sizeof(uint16_t) * sampleSpecSrc.getChannelCount());
 
     uint16_t *dstBuf = NULL;
-    uint32_t dstFrames = 0;
+    size_t dstFrames = 0;
     EXPECT_EQ(0, audioConversion.convert(sourceBuf, reinterpret_cast<void **>(&dstBuf),
                                          inputFrames, &dstFrames));
 
@@ -730,10 +730,10 @@ TEST(AudioConversion, memoryTest)
         10, 20, 5, 1, 3, 8, 12, 15, 10, 20, 5, 1, 3, 8, 12, 15,
         10, 20, 5, 1, 3, 8, 12, 15, 10, 20, 5, 1, 3, 8, 12, 15
     };
-    const uint32_t inputFrames =
+    const size_t inputFrames =
         sizeof(sourceBuf) / (sizeof(uint16_t) * sampleSpecSrc.getChannelCount());
 
-    uint32_t dstFrames = 0;
+    size_t dstFrames = 0;
     EXPECT_EQ(0, audioConversion->convert(sourceBuf,
                                           reinterpret_cast<void **>(&dstBuf),
                                           inputFrames, &dstFrames));
@@ -761,9 +761,9 @@ TEST(AudioConversion, frameExactApi)
         10, 20, 5, 1, 3, 8, 12, 15, 10, 20, 5, 1, 3, 8, 12, 15, 10, 20,
         10, 20, 5, 1, 3, 8, 12, 15, 10, 20, 5, 1, 3, 8, 12, 15, 10, 20
     };
-    const uint32_t inputFrames = sizeof(sourceBuf) /
+    const size_t inputFrames = sizeof(sourceBuf) /
                                  (sizeof(uint16_t) * sampleSpecSrc.getChannelCount());
-    const uint32_t expectedDstFrames = AudioUtils::convertSrcToDstInFrames(inputFrames,
+    const size_t expectedDstFrames = AudioUtils::convertSrcToDstInFrames(inputFrames,
                                                                            sampleSpecSrc,
                                                                            sampleSpecDst);
